requirements.txt
tensorflow==2.6
numpy==1.19.5
Pillow==7.2
